+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/jaxb/statement/dml/DeleteStatementTestCase.java
@@ -15,13 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.integrate.jaxb.statement.impl;
+package org.apache.shardingsphere.sql.parser.integrate.jaxb.statement.dml;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.insert.ExpectedInsertColumnsClause;
-import org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.insert.ExpectedInsertValuesClause;
-import org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.set.ExpectedSetClause;
+import org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.predicate.ExpectedWhereClause;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.segment.impl.table.ExpectedTable;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.statement.SQLParserTestCase;
 
@@ -31,24 +29,18 @@ import java.util.LinkedList;
 import java.util.List;
 
 /**
- * Insert statement test case.
+ * Delete statement test case.
  * 
  * @author zhangliang 
  */
 @Getter
 @Setter
-public final class InsertStatementTestCase extends SQLParserTestCase {
+public final class DeleteStatementTestCase extends SQLParserTestCase {
     
     @XmlElementWrapper
     @XmlElement(name = "table")
     private final List<ExpectedTable> tables = new LinkedList<>();
     
-    @XmlElement(name = "columns")
-    private ExpectedInsertColumnsClause insertColumnsClause;
-    
-    @XmlElement(name = "values")
-    private ExpectedInsertValuesClause insertValuesClause;
-    
-    @XmlElement(name = "set")
-    private ExpectedSetClause setClause;
+    @XmlElement(name = "where")
+    private ExpectedWhereClause whereClause;
 }
